third_party/blink/web_tests/external/wpt/css/css-fonts/inheritance-expected.txt

This is a testharness.js-based test.
[FAIL] Property font-synthesis has initial value weight style small-caps position
  assert_equals: expected "weight style small-caps position" but got "weight style small-caps"
Harness: the test ran to completion.
